新增了ai客服功能，提升用户体验和服务效率。

+++ hm-dianping/hm-dianping/src/main/java/com/hmdp/service/impl/FollowServiceImpl.java
@@ -13,7 +13,7 @@ import com.hmdp.utils.UserHolder;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
@@ -76,7 +76,7 @@ public class FollowServiceImpl extends ServiceImpl<FollowMapper, Follow> impleme
         //获取当前用户id
         Long userId = UserHolder.getUser().getId();
         //查询是否关注
-        Integer count = query().eq("user_id", userId).eq("follow_user_id", followUserId).count();
+        long count = query().eq("user_id", userId).eq("follow_user_id", followUserId).count();
         return Result.ok(count > 0);
     }
